fix: solve problem delete role

See merge request 123milhas/finances/api!459

+++ app/Http/Controllers/RoleController.php
@@ -2,9 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\DeleteRoleRequest;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\PutRoleRequest;
+use App\Models\ApprovalFlow;
+use App\Models\ApprovalFlowSupply;
+use App\Models\User;
 use App\Services\RoleService as RoleService;
 
 class RoleController extends Controller
@@ -40,6 +44,19 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
+        if(ApprovalFlow::where('role_id', $id)->exists() OR ApprovalFlowSupply::where('role_id', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este perfil está sendo utilizado no fluxo de aprovação é necessário que remova antes de apagar.'
+            ], 422);
+        }
+        if(User::where('role_id', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este prfil está sendo utilizado por usuário(s) no sistema é necessário que remova antes de apagar.'
+            ], 422);
+        }
+
         $role = $this->roleService->deleteRole($id);
         return response('');
     }
